pyproject.toml
[tool.poetry]
name = "djtest"
version = "0.1.0"
description = ""
authors = ["Ghazy <dianghazy@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
firebase-admin = "^6.2.0"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
